+++ packages/liveblocks-core/src/store.ts
@@ -115,8 +115,9 @@ type UpdateNotificationSettingsOptimisticUpdate = {
 };
 
 type QueryState =
-  | { isLoading: true; error?: never }
-  | { isLoading: false; error?: Error };
+  | { isLoading: true; error?: never } // pending
+  | { isLoading: false; error?: never } // success
+  | { isLoading: false; error: Error }; // error
 
 export type CacheState<M extends BaseMetadata> = {
   /**