+++ packages/liveblocks-react/src/liveblocks.tsx
@@ -10,13 +10,18 @@ import type {
   ClientOptions,
   DM,
   DU,
-  InboxNotificationData,
-  InboxNotificationDeleteInfo,
   OpaqueClient,
   PrivateClientApi,
-  ThreadDeleteInfo,
 } from "@liveblocks/core";
-import { createClient, kInternal, makePoller, raise } from "@liveblocks/core";
+import {
+  assert,
+  createClient,
+  kInternal,
+  makePoller,
+  memoizeOnSuccess,
+  raise,
+  shallow,
+} from "@liveblocks/core";
 import { nanoid } from "nanoid";
 import type { PropsWithChildren } from "react";
 import React, {
@@ -30,17 +35,16 @@ import { useSyncExternalStore } from "use-sync-external-store/shim/index.js";
 import { useSyncExternalStoreWithSelector } from "use-sync-external-store/shim/with-selector.js";
 
 import { selectedInboxNotifications } from "./comments/lib/selected-inbox-notifications";
-import { retryError } from "./lib/retry-error";
+import { autoRetry } from "./lib/retry-error";
 import { useInitial, useInitialUnlessFunction } from "./lib/use-initial";
+import { use } from "./lib/use-polyfill";
 import type {
   InboxNotificationsState,
-  InboxNotificationsStateSuccess,
   LiveblocksContextBundle,
   RoomInfoState,
   RoomInfoStateSuccess,
   SharedContextBundle,
   UnreadInboxNotificationsCountState,
-  UnreadInboxNotificationsCountStateSuccess,
   UserState,
   UserStateSuccess,
 } from "./types";
@@ -99,15 +103,6 @@ function selectorFor_useInboxNotifications(
   };
 }
 
-function selectorFor_useInboxNotificationsSuspense(
-  state: CacheState<BaseMetadata>
-): InboxNotificationsStateSuccess {
-  return {
-    inboxNotifications: selectedInboxNotifications(state),
-    isLoading: false,
-  };
-}
-
 function selectUnreadInboxNotificationsCount(state: CacheState<BaseMetadata>) {
   let count = 0;
 
@@ -141,15 +136,6 @@ function selectorFor_useUnreadInboxNotificationsCount(
     };
   }
 
-  return {
-    isLoading: false,
-    count: selectUnreadInboxNotificationsCount(state),
-  };
-}
-
-function selectorFor_useUnreadInboxNotificationsCountSuspense(
-  state: CacheState<BaseMetadata>
-): UnreadInboxNotificationsCountStateSuccess {
   return {
     isLoading: false,
     count: selectUnreadInboxNotificationsCount(state),
@@ -190,130 +176,120 @@ function makeExtrasForClient<U extends BaseUserMeta, M extends BaseMetadata>(
   const store = internals.cacheStore;
   const notifications = internals.notifications;
 
-  let fetchInboxNotifications$: Promise<{
-    inboxNotifications: InboxNotificationData[];
-    threads: ThreadData<M>[];
-    deletedThreads: ThreadDeleteInfo[];
-    deletedInboxNotifications: InboxNotificationDeleteInfo[];
-    meta: {
-      requestedAt: Date;
-    };
-  }> | null = null;
-
   let lastRequestedAt: Date | undefined;
 
-  const poller = makePoller(() =>
-    notifications.getInboxNotifications({ since: lastRequestedAt }).then(
-      (result) => {
-        lastRequestedAt = result.meta.requestedAt;
+  /**
+   * Performs one network fetch, and updates the store and last requested at
+   * date if successful. If unsuccessful, will throw.
+   */
+  async function updateInboxNotifications() {
+    const since =
+      lastRequestedAt !== undefined ? { since: lastRequestedAt } : undefined;
 
-        store.updateThreadsAndNotifications(
-          result.threads,
-          result.inboxNotifications,
-          result.deletedThreads,
-          result.deletedInboxNotifications,
-          INBOX_NOTIFICATIONS_QUERY
-        );
-      },
-      () => {
-        // TODO: Error handling
-      }
-    )
-  );
+    const result = await notifications.getInboxNotifications(since);
 
-  async function fetchInboxNotifications(
-    { retryCount }: { retryCount: number } = { retryCount: 0 }
-  ) {
-    if (fetchInboxNotifications$ !== null) {
-      return fetchInboxNotifications$;
+    store.updateThreadsAndNotifications(
+      result.threads,
+      result.inboxNotifications,
+      result.deletedThreads,
+      result.deletedInboxNotifications,
+      INBOX_NOTIFICATIONS_QUERY
+    );
+
+    /**
+     * We set the `lastRequestedAt` to the timestamp returned by the current request if:
+     * 1. The `lastRequestedAt` has not been set
+     * OR
+     * 2. The current `lastRequestedAt` is older than the timestamp returned by the current request
+     */
+    if (
+      lastRequestedAt === undefined ||
+      lastRequestedAt < result.meta.requestedAt
+    ) {
+      lastRequestedAt = result.meta.requestedAt;
     }
+  }
 
+  let pollerSubscribers = 0;
+  const poller = makePoller(async () => {
+    return waitUntilInboxNotificationsLoaded()
+      .then(updateInboxNotifications)
+      .catch(() => {
+        // When polling, we don't want to throw errors, ever
+        // XXX Maybe issue console warnings here though?
+      });
+  });
+
+  /**
+   * Will trigger an initial fetch of inbox notifications if this hasn't
+   * already happened. Will resolve once there is initial data. Will retry
+   * a few times automatically in case fetching fails, with incremental backoff
+   * delays. Will throw eventually only if all retries fail.
+   */
+  const waitUntilInboxNotificationsLoaded = memoizeOnSuccess(async () => {
     store.setQueryState(INBOX_NOTIFICATIONS_QUERY, {
       isLoading: true,
     });
 
     try {
-      fetchInboxNotifications$ = notifications.getInboxNotifications();
-
-      const result = await fetchInboxNotifications$;
-
-      store.updateThreadsAndNotifications(
-        result.threads,
-        result.inboxNotifications,
-        result.deletedThreads,
-        result.deletedInboxNotifications,
-        INBOX_NOTIFICATIONS_QUERY
+      await autoRetry(
+        () => updateInboxNotifications(),
+        5,
+        // XXX: We're currently doing wait as high as 40000, 80000 here, but...
+        // do we really wait until over a minute? Seems too long to me. Maybe
+        // instead we want to try a bit more often, and with a bit less waiting
+        // time?
+        // XXX: Proposal: change the array below to [5000, 5000, 10000, 10000, 15000, 15000] (= total timeout of 1 minute)
+        [5000, 10000, 20000, 40000, 80000]
       );
-
-      /**
-       * We set the `lastRequestedAt` to the timestamp returned by the current request if:
-       * 1. The `lastRequestedAt` has not been set
-       * OR
-       * 2. The current `lastRequestedAt` is older than the timestamp returned by the current request
-       */
-      if (
-        lastRequestedAt === undefined ||
-        lastRequestedAt > result.meta.requestedAt
-      ) {
-        lastRequestedAt = result.meta.requestedAt;
-      }
-
-      poller.start(POLLING_INTERVAL);
-    } catch (er) {
-      fetchInboxNotifications$ = null;
-
-      // Retry the action using the exponential backoff algorithm
-      retryError(() => {
-        void fetchInboxNotifications({
-          retryCount: retryCount + 1,
-        });
-      }, retryCount);
-
+    } catch (err) {
+      // Store the error in the cache as a side-effect, for non-Suspense
       store.setQueryState(INBOX_NOTIFICATIONS_QUERY, {
         isLoading: false,
-        error: er as Error,
+        error: err as Error,
       });
+
+      // Rethrow it for Suspense, where this promise must fail
+      throw err;
     }
-    return;
-  }
+  });
 
-  let inboxNotificationsSubscribers = 0;
-
-  function useSubscribeToInboxNotificationsEffect(options?: {
-    autoFetch: boolean;
-  }) {
-    const autoFetch = useInitial(options?.autoFetch ?? true);
+  /**
+   * Enables polling for inbox notifications when the component mounts. Stops
+   * polling on unmount.
+   *
+   * Safe to be called multiple times from different components. The first
+   * component to mount starts the polling. The last component to unmount stops
+   * the polling.
+   */
+  function useEnableInboxNotificationsPolling() {
     useEffect(() => {
-      if (autoFetch) {
-        void fetchInboxNotifications();
-      }
-
       // Increment
-      inboxNotificationsSubscribers++;
+      pollerSubscribers++;
       poller.start(POLLING_INTERVAL);
 
       return () => {
         // Decrement
-        if (inboxNotificationsSubscribers <= 0) {
+        if (pollerSubscribers <= 0) {
           console.warn(
             `Internal unexpected behavior. Cannot decrease subscriber count for query "${INBOX_NOTIFICATIONS_QUERY}"`
           );
           return;
         }
 
-        inboxNotificationsSubscribers--;
-        if (inboxNotificationsSubscribers <= 0) {
+        pollerSubscribers--;
+        if (pollerSubscribers <= 0) {
           poller.stop();
         }
       };
-    }, [autoFetch]);
+    }, []);
   }
 
   return {
     store,
     notifications,
-    fetchInboxNotifications,
-    useSubscribeToInboxNotificationsEffect,
+    useEnableInboxNotificationsPolling,
+    waitUntilInboxNotificationsLoaded,
   };
 }
 
@@ -378,79 +354,59 @@ function makeLiveblocksContextBundle<
 }
 
 function useInboxNotifications_withClient(client: OpaqueClient) {
-  const { store, useSubscribeToInboxNotificationsEffect } =
+  const { store, useEnableInboxNotificationsPolling } =
     getExtrasForClient(client);
 
-  useSubscribeToInboxNotificationsEffect();
+  useEnableInboxNotificationsPolling();
   return useSyncExternalStoreWithSelector(
     store.subscribe,
     store.get,
     store.get,
-    selectorFor_useInboxNotifications
+    selectorFor_useInboxNotifications,
+    shallow
   );
 }
 
 function useInboxNotificationsSuspense_withClient(client: OpaqueClient) {
-  const {
-    store,
-    fetchInboxNotifications,
-    useSubscribeToInboxNotificationsEffect,
-  } = getExtrasForClient(client);
+  const { waitUntilInboxNotificationsLoaded } = getExtrasForClient(client);
 
-  const query = store.get().queries[INBOX_NOTIFICATIONS_QUERY];
+  // Suspend until there are at least some inbox notifications
+  use(waitUntilInboxNotificationsLoaded());
 
-  if (query === undefined || query.isLoading) {
-    throw fetchInboxNotifications();
-  }
-
-  if (query.error !== undefined) {
-    throw query.error;
-  }
-
-  useSubscribeToInboxNotificationsEffect({ autoFetch: false });
-  return useSyncExternalStoreWithSelector(
-    store.subscribe,
-    store.get,
-    store.get,
-    selectorFor_useInboxNotificationsSuspense
-  );
+  // We're in a Suspense world here, and as such, the useInboxNotifications()
+  // hook is expected to only return success results when we're here.
+  const result = useInboxNotifications_withClient(client);
+  assert(!result.error, "Did not expect error");
+  assert(!result.isLoading, "Did not expect loading");
+  return result;
 }
 
 function useUnreadInboxNotificationsCount_withClient(client: OpaqueClient) {
-  const { store, useSubscribeToInboxNotificationsEffect } =
+  const { store, useEnableInboxNotificationsPolling } =
     getExtrasForClient(client);
 
-  useSubscribeToInboxNotificationsEffect();
+  useEnableInboxNotificationsPolling();
   return useSyncExternalStoreWithSelector(
     store.subscribe,
     store.get,
     store.get,
-    selectorFor_useUnreadInboxNotificationsCount
+    selectorFor_useUnreadInboxNotificationsCount,
+    shallow
   );
 }
 
 function useUnreadInboxNotificationsCountSuspense_withClient(
   client: OpaqueClient
 ) {
-  const {
-    store,
-    fetchInboxNotifications,
-    useSubscribeToInboxNotificationsEffect,
-  } = getExtrasForClient(client);
+  const { waitUntilInboxNotificationsLoaded } = getExtrasForClient(client);
 
-  const query = store.get().queries[INBOX_NOTIFICATIONS_QUERY];
+  // Suspend until there are at least some inbox notifications
+  use(waitUntilInboxNotificationsLoaded());
 
-  if (query === undefined || query.isLoading) {
-    throw fetchInboxNotifications();
-  }
-
-  useSubscribeToInboxNotificationsEffect({ autoFetch: false });
-  return useSyncExternalStoreWithSelector(
-    store.subscribe,
-    store.get,
-    store.get,
-    selectorFor_useUnreadInboxNotificationsCountSuspense
-  );
+  const result = useUnreadInboxNotificationsCount_withClient(client);
+  assert(!result.isLoading, "Did not expect loading");
+  assert(!result.error, "Did not expect error");
+  return result;
 }
 
 function useMarkInboxNotificationAsRead_withClient(client: OpaqueClient) {