+++ packages/liveblocks-react/src/types.ts
@@ -171,12 +171,6 @@ export type InboxNotificationsStateLoading = {
   error?: never;
 };
 
-export type InboxNotificationsStateResolved = {
-  isLoading: false;
-  inboxNotifications: InboxNotificationData[];
-  error?: Error;
-};
-
 export type InboxNotificationsStateSuccess = {
   isLoading: false;
   inboxNotifications: InboxNotificationData[];
@@ -191,7 +185,7 @@ export type InboxNotificationsStateError = {
 
 export type InboxNotificationsState =
   | InboxNotificationsStateLoading
-  | InboxNotificationsStateResolved
+  | InboxNotificationsStateSuccess
   | InboxNotificationsStateError;
 
 export type UnreadInboxNotificationsCountStateLoading = {