+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - The `useClient()` hook is now also available for users of
   `createRoomContext()` and/or `createLiveblocksContext()`
+- Fix bug where all subsequent fetches of inbox notifications were not
+  incremental.
+- Fix: avoid unnecessary re-renders if inbox notifications haven't changed
 
 ### `@liveblocks/react-ui`